MainVC - commented old tableview updates, added filterTodaysHabit()
commented all the codes that updates tableView based on changes in RMO_Habits.
Included filterTodaysHabit() into didCreateNewHabit() so it filters. this solved the error of " Terminating app due to uncaught exception 'NSInternalInconsistencyException', reason: 'attempt to insert row 2 into section 0, but there are only 2 rows in section 0 after the update' ".
STILL Errors out in AllHabitsVC, and when updating while Searching.