MainVC - initHabits() updated. HabitDetailVC - moved updating date function to MainVC
MainVC -
initHabit() updated. Now date of repeated habits get updated when the app is ran. all the funcitonality should work. further testing is required. Probably better if we move this to userDefault area.
HabitDetailVC -
deleted code that does that same thing as the one mentioned above. repeated habits date used to get updated when the habit was deleted, but now it gets updated when app starts