HabitDetailVC - only change rate when success. NewHabitVC, MainVC - rate auto generated as "Fail" when first created or ran app
HabitDetailVC - only change rate when success. fail is automatically calculated when app is ran OR when the habit is initially created.
NewHabitVC, MainVC - rate auto generated as "Fail" when first created or ran app.

this solves the problem of user opening the app but not doing anything and leaving gaps in dataset used for graph