MainVC - appEnterForeground added. HabitDetailNoReVC, HabitDetailVC, NewHabitVC - title text to "No Title" when empty
MainVC - appEnterForeground added.
Now when the user doesn't closes the app but simply jumps back on the app while the app is still live in the background, it updates habits + date + graph if its the next day.
HabitDetailNoReVC, HabitDetailVC, NewHabitVC - title text to "No Title" when empty so the constraints for Desc and Date don't move when title is empty.
Updated "description of your new habit" to "please add description" when user didn't add desc