RMO_Count - added finalPercent, MainVC - need to update the final percent when the day is over (NEED FIX), ProgressVC -  graph updated to show progress of all habits in LineChart
RMO_Count - added finalPercent, this will contain the % of the days success rate.
MainVC - need to save the final percent when the day is over in Count_Realm(NEED FIX) for progressVC to use.
ProgressVC -  graph updated to show progress of all habits in LineChart. need a lot of fixing