RMO_Habit - todaysResult added. HabitDetailVC - todayResult changes based on which button pressed. success rate gets deleted if todayResult == 1 when deleting habit. MainVC - weekly logic fixed
RMO_Habit - todaysResult added - this is used later to accurately calculated % completed.
HabitDetailVC - todayResult changes based on which button pressed. if success was pressed, we should -1 from success so it doesn't display in the success %. otherwise if 8 habits were succeeded, but we delete one and now the total is 7, the new success rate will be over 100%
MainVC - weekly logic fixed again. NOW a lot has changed. the function runs when dayDifference > 0 , and adds 7 days to dayDifference and floors down. so if 1 day have passed, it will + 7 and round down so multiplesOfSeven would be 1. BEFORE, the dates would change ONLY if the user used the app on the 7th from the original habit, but now with this logic, the user can access whenever and it will calculate 0% for missed days and assign future date. then only on the day when daydifferene%7 == 0, which is 1 week away from the habit date, count increases by 1. if daydifference is -, the app doesn't do anything